# Generate keys outside benchmark loop