The QPS results are tainted due to generating the k/v objects inside the hot benchmark loop.

This PR moves the object generation outside the loop.

**Before**
```
itemSize=500000, workloads=7500000, cacheSize=0.06%, zipf's alpha=0.99, concurrency=1

      CACHE      | HITRATE | MEMORY  |   QPS   |  HITS   | MISSES   
-----------------+---------+---------+---------+---------+----------
  otter          | 44.16%  | 0.72MiB | 3598848 | 3311718 | 4188282  
  sieve          | 43.95%  | 0.06MiB | 4201681 | 3296584 | 4203416  
  tinylfu        | 43.71%  | 0.06MiB | 4136790 | 3278140 | 4221860  
  s3-fifo        | 43.54%  | 0.11MiB | 2618715 | 3265799 | 4234201  
  slru           | 42.87%  | 0.06MiB | 3588517 | 3215058 | 4284942  
  s4lru          | 42.52%  | 0.06MiB | 4980080 | 3189215 | 4310785  
  two-queue      | 41.89%  | 0.10MiB | 2593361 | 3141510 | 4358490  
  clock          | 33.33%  | 0.05MiB | 3978780 | 2499452 | 5000548  
  lru-hashicorp  | 32.57%  | 0.02MiB | 3778338 | 2442520 | 5057480  
  lru-groupcache | 32.57%  | 0.06MiB | 3890041 | 2442520 | 5057480  
```

**After**
```
itemSize=500000, workloads=7500000, cacheSize=0.06%, zipf's alpha=0.99, concurrency=1

      CACHE      | HITRATE | MEMORY  |   QPS    |  HITS   | MISSES   
-----------------+---------+---------+----------+---------+----------
  otter          | 44.42%  | 0.85MiB |  6019262 | 3331547 | 4168453  
  sieve          | 43.95%  | 0.06MiB |  7637475 | 3296584 | 4203416  
  tinylfu        | 43.70%  | 0.06MiB |  7953340 | 3277571 | 4222429  
  s3-fifo        | 43.54%  | 0.11MiB |  3810976 | 3265799 | 4234201  
  slru           | 42.87%  | 0.06MiB |  6702413 | 3215058 | 4284942  
  s4lru          | 42.52%  | 0.06MiB | 10373444 | 3189215 | 4310785  
  two-queue      | 41.89%  | 0.10MiB |  3646087 | 3141510 | 4358490  
  clock          | 33.33%  | 0.05MiB |  6893382 | 2499452 | 5000548  
  lru-hashicorp  | 32.57%  | 0.05MiB |  6732496 | 2442520 | 5057480  
  lru-groupcache | 32.57%  | 0.06MiB |  6843066 | 2442520 | 5057480  
```